# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"defaultFisrtName": {
				"type": "string",
				"example": "firstName",
				"description": "Field for name"
			},
			"defaultLastName": {
				"type": "string",
				"example": "lastName",
				"description": "Field for lastName"
			},
			"defaultEmail": {
				"type": "string",
				"example": "ftchienguen@gmail.com",
				"description": "Field for name"
			},
			"UserType.userUpdateFields": {
				"properties": {
					"firstName": {
						"$ref": "#/components/schemas/defaultFisrtName"
					},
					"lastName": {
						"$ref": "#/components/schemas/defaultLastName"
					},
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					}
				},
				"required": [
					"firstName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType.loginFields": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType.forgotPasswordFields": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"defaultOtp": {
				"type": "number",
				"format": "double",
				"example": 1111,
				"description": "Field for otp"
			},
			"VERIF_TYPE": {
				"enum": [
					"login",
					"forgot_pass"
				],
				"type": "string"
			},
			"UserType.verifyOtp": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					},
					"otp": {
						"$ref": "#/components/schemas/defaultOtp"
					},
					"verificationType": {
						"$ref": "#/components/schemas/VERIF_TYPE"
					}
				},
				"required": [
					"email",
					"otp",
					"verificationType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType.resendOtp": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType.changePasswordFields": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					},
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType.userCreateFields": {
				"properties": {
					"firstName": {
						"$ref": "#/components/schemas/defaultFisrtName"
					},
					"lastName": {
						"$ref": "#/components/schemas/defaultLastName"
					},
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Jwt": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header",
				"scopes": {
					"root": "can do",
					"admin": "can do",
					"user": "can do"
				}
			}
		}
	},
	"info": {
		"title": "authenticationservice",
		"version": "1.0.0",
		"description": "Auth service api",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [
					{
						"Jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/edit": {
			"put": {
				"operationId": "Edit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.userUpdateFields"
							}
						}
					}
				}
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.loginFields"
							}
						}
					}
				}
			}
		},
		"/forgot_password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.forgotPasswordFields"
							}
						}
					}
				}
			}
		},
		"/verify-otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.verifyOtp"
							}
						}
					}
				}
			}
		},
		"/resent-otp": {
			"post": {
				"operationId": "Resendotp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.resendOtp"
							}
						}
					}
				}
			}
		},
		"/change_password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [
					{
						"Jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.changePasswordFields"
							}
						}
					}
				}
			}
		},
		"/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.userCreateFields"
							}
						}
					}
				}
			}
		},
		"/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Controller"
				],
				"security": [
					{
						"Jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3000/"
		}
	]
}